moved get_legacy_frontiers to independent mod
